pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.14",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
